Added shop and player utilities. Added user interface for remaining time and money. Minor changes. Renamed Vehicle super class as Weapon.
