+++ Assets/Scripts/PlayerStats.cs
@@ -1,0 +1,21 @@
+﻿using UnityEngine;
+
+public class PlayerStats : MonoBehaviour
+{
+    public static int money;
+    public int startMoney = 100;
+
+//    public int life = 10;
+//    private float totalDamage;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        money = startMoney;
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+    }
+}
