+++ Assets/Scripts/Tank.cs
@@ -1,4 +1,4 @@
-﻿public class Tank : Vehicle
+﻿public class Tank : Weapon
 {
     // Start is called before the first frame update
 }
