--- Assets/Scripts/Vehicle.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: fb4c83a02b334549b6875686125d4564
-timeCreated: 1552415558
