+++ Assets/Scripts/Weapon.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 8d25cff54e64642059e03a595c867dad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: fb4c83a02b334549b6875686125d4564
+timeCreated: 1552415558
